perf: reduce runtime and test startup overhead
