+++ extensions/slack/src/directory-config.ts
@@ -2,7 +2,7 @@ import {
   listInspectedDirectoryEntriesFromSources,
   type DirectoryConfigParams,
 } from "openclaw/plugin-sdk/directory-runtime";
-import { inspectSlackAccount, type InspectedSlackAccount } from "../api.js";
+import { inspectSlackAccount, type InspectedSlackAccount } from "./account-inspect.js";
 import { parseSlackTarget } from "./targets.js";
 
 export async function listSlackDirectoryPeersFromConfig(params: DirectoryConfigParams) {
