+++ extensions/whatsapp/auth-presence.ts
@@ -1,0 +1,1 @@
+export { hasAnyWhatsAppAuth } from "./src/accounts.js";
