+++ src/cron/isolated-agent/run.sandbox-config-preserved.test.ts
@@ -1,55 +1,45 @@
-import { afterEach, beforeEach, describe, expect, it } from "vitest";
-import {
-  clearFastTestEnv,
-  loadRunCronIsolatedAgentTurn,
-  resolveAgentConfigMock,
-  resetRunCronIsolatedAgentTurnHarness,
-  restoreFastTestEnv,
-  runWithModelFallbackMock,
-} from "./run.test-harness.js";
+import { describe, expect, it } from "vitest";
+import { resolveSandboxConfigForAgent } from "../../agents/sandbox/config.js";
+import { buildCronAgentDefaultsConfig } from "./run-config.js";
 
-const runCronIsolatedAgentTurn = await loadRunCronIsolatedAgentTurn();
-
-function makeJob(overrides?: Record<string, unknown>) {
+function makeCfg() {
   return {
-    id: "sandbox-test-job",
-    name: "Sandbox Test",
-    schedule: { kind: "cron", expr: "0 9 * * *", tz: "UTC" },
-    sessionTarget: "isolated",
-    payload: { kind: "agentTurn", message: "test" },
-    ...overrides,
-  } as never;
-}
-
-function makeParams(overrides?: Record<string, unknown>) {
-  return {
-    cfg: {
-      agents: {
-        defaults: {
-          sandbox: {
-            mode: "all" as const,
-            workspaceAccess: "rw" as const,
-            docker: {
-              network: "none",
-              dangerouslyAllowContainerNamespaceJoin: true,
-              dangerouslyAllowExternalBindSources: true,
-            },
-            browser: {
-              enabled: true,
-              autoStart: false,
-            },
-            prune: {
-              maxAgeDays: 7,
-            },
+    agents: {
+      defaults: {
+        sandbox: {
+          mode: "all" as const,
+          workspaceAccess: "rw" as const,
+          docker: {
+            network: "none",
+            dangerouslyAllowContainerNamespaceJoin: true,
+            dangerouslyAllowExternalBindSources: true,
+          },
+          browser: {
+            enabled: true,
+            autoStart: false,
+          },
+          prune: {
+            maxAgeDays: 7,
           },
         },
       },
     },
-    deps: {} as never,
-    job: makeJob(),
-    message: "test",
-    sessionKey: "cron:sandbox-test",
-    ...overrides,
+  };
+}
+
+function buildRunCfg(agentId: string, agentConfigOverride?: Record<string, unknown>) {
+  const cfg = makeCfg();
+  const agentDefaults = buildCronAgentDefaultsConfig({
+    defaults: cfg.agents.defaults,
+    agentConfigOverride: agentConfigOverride as never,
+  });
+  return {
+    ...cfg,
+    agents: {
+      ...cfg.agents,
+      defaults: agentDefaults,
+      list: [{ id: agentId, ...agentConfigOverride }],
+    },
   };
 }
 
@@ -79,35 +69,23 @@ function expectDefaultSandboxPreserved(
 }
 
 describe("runCronIsolatedAgentTurn sandbox config preserved", () => {
-  let previousFastTestEnv: string | undefined;
-
-  beforeEach(() => {
-    previousFastTestEnv = clearFastTestEnv();
-    resetRunCronIsolatedAgentTurnHarness();
-  });
-
-  afterEach(() => {
-    restoreFastTestEnv(previousFastTestEnv);
-  });
-
   it("preserves default sandbox config when agent entry omits sandbox", async () => {
-    resolveAgentConfigMock.mockReturnValue({
+    const runCfg = buildRunCfg("worker", {
       name: "worker",
       workspace: "/tmp/custom-workspace",
       sandbox: undefined,
       heartbeat: undefined,
       tools: undefined,
     });
-
-    await runCronIsolatedAgentTurn(makeParams({ agentId: "worker" }));
-
-    expect(runWithModelFallbackMock).toHaveBeenCalledTimes(1);
-    const runCfg = runWithModelFallbackMock.mock.calls[0]?.[0]?.cfg;
     expectDefaultSandboxPreserved(runCfg);
+    expect(resolveSandboxConfigForAgent(runCfg, "worker")).toMatchObject({
+      mode: "all",
+      workspaceAccess: "rw",
+    });
   });
 
   it("keeps global sandbox defaults when agent override is partial", async () => {
-    resolveAgentConfigMock.mockReturnValue({
+    const runCfg = buildRunCfg("specialist", {
       sandbox: {
         docker: {
           image: "ghcr.io/openclaw/sandbox:custom",
@@ -120,12 +98,6 @@ describe("runCronIsolatedAgentTurn sandbox config preserved", () => {
         },
       },
     });
-
-    await runCronIsolatedAgentTurn(makeParams({ agentId: "specialist" }));
-
-    expect(runWithModelFallbackMock).toHaveBeenCalledTimes(1);
-    const runCfg = runWithModelFallbackMock.mock.calls[0]?.[0]?.cfg;
-    const { resolveSandboxConfigForAgent } = await import("../../agents/sandbox/config.js");
     const resolvedSandbox = resolveSandboxConfigForAgent(runCfg, "specialist");
 
     expectDefaultSandboxPreserved(runCfg);
