+++ src/memory/test-manager-helpers.ts
@@ -4,12 +4,14 @@ import type { MemoryIndexManager } from "./index.js";
 export async function getRequiredMemoryIndexManager(params: {
   cfg: OpenClawConfig;
   agentId?: string;
+  purpose?: "default" | "status";
 }): Promise<MemoryIndexManager> {
   await import("./embedding.test-mocks.js");
   const { getMemorySearchManager } = await import("./index.js");
   const result = await getMemorySearchManager({
     cfg: params.cfg,
     agentId: params.agentId ?? "main",
+    purpose: params.purpose,
   });
   if (!result.manager) {
     throw new Error("manager missing");
