+++ src/plugin-sdk/index.test.ts
@@ -2,10 +2,49 @@ import fs from "node:fs/promises";
 import path from "node:path";
 import { describe, expect, it } from "vitest";
 import { buildPluginSdkPackageExports } from "./entrypoints.js";
-import * as sdk from "./index.js";
+
+async function collectRuntimeExports(filePath: string, seen = new Set<string>()) {
+  const normalizedPath = path.resolve(filePath);
+  if (seen.has(normalizedPath)) {
+    return new Set<string>();
+  }
+  seen.add(normalizedPath);
+
+  const source = await fs.readFile(normalizedPath, "utf8");
+  const exportNames = new Set<string>();
+
+  for (const match of source.matchAll(/export\s+(?!type\b)\{([\s\S]*?)\}\s+from\s+"([^"]+)";/g)) {
+    const names = match[1]
+      .split(",")
+      .map((part) => part.trim())
+      .filter(Boolean)
+      .map((part) => part.split(/\s+as\s+/).at(-1) ?? part);
+    for (const name of names) {
+      exportNames.add(name);
+    }
+  }
+
+  for (const match of source.matchAll(/export\s+\*\s+from\s+"([^"]+)";/g)) {
+    const specifier = match[1];
+    if (!specifier.startsWith(".")) {
+      continue;
+    }
+    const nestedPath = path.resolve(
+      path.dirname(normalizedPath),
+      specifier.replace(/\.js$/, ".ts"),
+    );
+    const nestedExports = await collectRuntimeExports(nestedPath, seen);
+    for (const name of nestedExports) {
+      exportNames.add(name);
+    }
+  }
+
+  return exportNames;
+}
 
 describe("plugin-sdk exports", () => {
-  it("does not expose runtime modules", () => {
+  it("does not expose runtime modules", async () => {
+    const runtimeExports = await collectRuntimeExports(path.join(import.meta.dirname, "index.ts"));
     const forbidden = [
       "chunkMarkdownText",
       "chunkText",
@@ -43,18 +82,21 @@ describe("plugin-sdk exports", () => {
     ];
 
     for (const key of forbidden) {
-      expect(Object.prototype.hasOwnProperty.call(sdk, key)).toBe(false);
+      expect(runtimeExports.has(key)).toBe(false);
     }
   });
 
-  it("keeps the root runtime surface intentionally small", () => {
-    expect(typeof sdk.emptyPluginConfigSchema).toBe("function");
-    expect(typeof sdk.delegateCompactionToRuntime).toBe("function");
-    expect(typeof sdk.onDiagnosticEvent).toBe("function");
-    expect(Object.prototype.hasOwnProperty.call(sdk, "resolveControlCommandGate")).toBe(false);
-    expect(Object.prototype.hasOwnProperty.call(sdk, "buildAgentSessionKey")).toBe(false);
-    expect(Object.prototype.hasOwnProperty.call(sdk, "isDangerousNameMatchingEnabled")).toBe(false);
-    expect(Object.prototype.hasOwnProperty.call(sdk, "emitDiagnosticEvent")).toBe(false);
+  it("keeps the root runtime surface intentionally small", async () => {
+    const runtimeExports = await collectRuntimeExports(path.join(import.meta.dirname, "index.ts"));
+    expect([...runtimeExports].toSorted()).toEqual([
+      "buildFalImageGenerationProvider",
+      "buildGoogleImageGenerationProvider",
+      "buildOpenAIImageGenerationProvider",
+      "delegateCompactionToRuntime",
+      "emptyPluginConfigSchema",
+      "onDiagnosticEvent",
+      "registerContextEngine",
+    ]);
   });
 
   it("keeps package.json plugin-sdk exports synced with the manifest", async () => {
